Implementing Firebase - Hailey
1. implemented firebase authentication
uses firebase and the login_page.html document to create users accounts and store them.

2. update expense_input.html
added some scripts to the body that check if a user is logged in before allowing access to the page. I added this on top of the edits previously made, so it should still be up to date.

3. edited the homepage
 changed the button to redirect to the login page instead of the expense_input.html page.

4. added a firebase-config file
this file is what allows our pages to use firebase